apps/api/requirements.txt
setuptools
hypercorn
fastapi
fastapi[standard]
httpx
sqlalchemy
sqlalchemy[asyncio]
psycopg
psycopg[binary]
langchain[llms]
langchain
langgraph
dependency-injector
alembic
aiobcrypt
pydantic
python-dotenv
redis
pyjwt[crypto]
passlib
pytz
black
langchain-openai
langgraph-checkpoint-postgres